requirements.txt
fastapi
uvicorn
langgraph
pandas
openai
pinecone-client
pdfplumber
python-dotenv
streamlit
requests
